src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file=src/requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via microsoft-kiota-authentication-azure
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.0
    # via fastmcp
azure-core==1.36.0
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.25.1
    # via msgraph-sdk
beartype==0.22.5
    # via
    #   py-key-value-aio
    #   py-key-value-shared
cachetools==6.2.1
    # via py-key-value-aio
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.2.1 ; sys_platform != 'emscripten'
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==45.0.3
    # via
    #   authlib
    #   azure-identity
    #   msal
    #   pyjwt
    #   secretstorage
cyclopts==4.2.3
    # via fastmcp
diskcache==5.6.3
    # via py-key-value-aio
distlib==0.4.0
    # via virtualenv
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.3
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.0
    # via fastmcp
fastmcp==2.13.0.2
    # via bot-conversation-python
filelock==3.20.0
    # via virtualenv
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
    #   microsoft-kiota-http
    #   msgraph-core
httpx-sse==0.4.0
    # via mcp
hyperframe==6.1.0
    # via h2
identify==2.6.15
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.3.0
    # via keyring
jeepney==0.9.0 ; sys_platform == 'linux'
    # via
    #   keyring
    #   secretstorage
jsonschema==4.25.1
    # via mcp
jsonschema-path==0.3.4
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.6.0
    # via py-key-value-aio
markdown-it-py==3.0.0
    # via rich
mcp==1.21.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
microsoft-kiota-abstractions==1.9.7
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-core
microsoft-kiota-authentication-azure==1.9.7
    # via msgraph-core
microsoft-kiota-http==1.9.7
    # via msgraph-core
microsoft-kiota-serialization-form==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-json==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-multipart==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.9.7
    # via msgraph-sdk
more-itertools==10.8.0
    # via
    #   jaraco-classes
    #   jaraco-functools
msal==1.34.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msgraph-core==1.3.8
    # via msgraph-sdk
msgraph-sdk==1.48.0
    # via bot-conversation-python
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nodeenv==1.9.1
    # via pre-commit
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.38.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.38.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
pathable==0.4.4
    # via jsonschema-path
pathvalidate==3.3.1
    # via py-key-value-aio
platformdirs==4.5.0
    # via
    #   fastmcp
    #   virtualenv
pre-commit==4.4.0
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
py-key-value-aio==0.2.8
    # via fastmcp
py-key-value-shared==0.2.8
    # via py-key-value-aio
pycparser==2.22 ; platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.4
    # via
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via
    #   mcp
    #   msal
pyperclip==1.11.0
    # via fastmcp
python-dotenv==1.2.1
    # via
    #   bot-conversation-python
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pywin32==311 ; sys_platform == 'win32'
    # via mcp
pywin32-ctypes==0.2.3 ; sys_platform == 'win32'
    # via keyring
pyyaml==6.0.3
    # via
    #   jsonschema-path
    #   pre-commit
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   azure-core
    #   bot-conversation-python
    #   jsonschema-path
    #   msal
rich==14.0.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
secretstorage==3.4.1 ; sys_platform == 'linux'
    # via keyring
sniffio==1.3.1
    # via anyio
sse-starlette==2.3.6
    # via mcp
starlette==0.37.2
    # via mcp
std-uritemplate==2.0.8
    # via microsoft-kiota-abstractions
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   azure-core
    #   azure-identity
    #   exceptiongroup
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   py-key-value-shared
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.38.0 ; sys_platform != 'emscripten'
    # via mcp
virtualenv==20.35.4
    # via pre-commit
websockets==15.0.1
    # via fastmcp
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
